% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matinput.R
\name{matinput}
\alias{matinput}
\title{Launch the MatSamp data entry app}
\usage{
matinput()
}
\value{
A Shiny app launched in the current R session.
}
\description{
Starts a Shiny app for entering sampling data with fields for date,
LFA, organization, and measurement.
}
